Raxol.Core.Runtime.ProcessStore (Raxol Core v2.4.0)

Replacement for Process dictionary usage. Provides a functional alternative using Agent for state storage.

Functions

child_spec(arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

clear()

Clears all values from the store.

delete(key)

Deletes a value from the store.

get(key, default \\ nil)

Gets a value from the store.

get_all()

Gets all values from the store.

get_and_update(key, fun)

Gets and updates a value atomically.

put(key, value)

Puts a value in the store.

start_link(initial_state \\ %{})

Starts a new process store.

update(key, default, fun)

Updates a value in the store using a function.